[component] Move internal/sharedcomponent to component/sharedcomponent
