+++ component/sharedcomponent/sharedcomponent_test.go
